[database] Moved album to its own file
